# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "llShop后端接口API",
        "title": "llShop-backend",
        "termsOfService": "https://github.com/mayooot/llShop-backend",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "y8ccqm.natappfree.cc",
    "basePath": "/api/v1",
    "paths": {
        "/phone": {
            "get": {
                "description": "前端传递11位手机号，后端随机生成4位验证码。存入Redis，并异步调用阿里云SMS服务发送验证码。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "获取验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/signup": {
            "post": {
                "description": "前端传递JSON类型对象，后端完成校验后注册新用户。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "验证码",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamSignUp"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.ParamSignUp": {
            "type": "object",
            "required": [
                "code",
                "password",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}